Add files via upload
Indian Premier League (IPL) Data Analysis by using Python & Power BI.
In this Data set we have two files one is Matches file and another one is Delivery File. We have to Analyze both file and by seeing the visualization As a sports analysts, we have to find out the most successful teams, players and factors contributing win or loss of a team. We have to suggest teams or players a company should endorse for its products.